Move all PlotParameters to plotting module
